Address a problem where output cannot be captured when all IO is redirected
It seems like we get an EOF too early because the OS does not expect
things to come from /dev/tty.
So, when everything is redirected, we just get an EOF because there is
no real source of information.
This commit is kind of a workaround. See the comment in process.py for details.